Initial windows support for agent (#3375)
