+++ pkg/agent/loadbalancer/utility_windows.go
@@ -1,0 +1,9 @@
+// +build windows
+
+package loadbalancer
+
+import "syscall"
+
+func reusePort(network, address string, conn syscall.RawConn) error {
+	return nil
+}
