+++ pkg/agent/syssetup/setup_windows.go
@@ -1,3 +1,7 @@
 package syssetup
 
-func Configure() {}
+import kubeproxyconfig "k8s.io/kubernetes/pkg/proxy/apis/config"
+
+func Configure(enableIPv6 bool, config *kubeproxyconfig.KubeProxyConntrackConfiguration) {
+
+}
