+++ pkg/cgroups/cgroups_windows.go
@@ -1,0 +1,11 @@
+// +build windows
+
+package cgroups
+
+func Validate() error {
+	return nil
+}
+
+func CheckCgroups() (kubeletRoot, runtimeRoot string, hasCFS, hasPIDs bool) {
+	return "", "", false, false
+}
