+++ pkg/cli/cmds/root.go
@@ -1,7 +1,6 @@
 package cmds
 
 import (
-	"errors"
 	"fmt"
 	"os"
 	"runtime"
@@ -20,8 +19,6 @@ var (
 		EnvVar:      version.ProgramUpper + "_DEBUG",
 	}
 )
-
-var ErrCommandNoArgs = errors.New("this command does not take any arguments")
 
 func init() {
 	// hack - force "file,dns" lookup order if go dns is used
