+++ pkg/containerd/utility_linux.go
@@ -1,0 +1,16 @@
+// +build linux
+
+package containerd
+
+import (
+	"github.com/containerd/containerd/snapshots/overlay"
+	fuseoverlayfs "github.com/containerd/fuse-overlayfs-snapshotter"
+)
+
+func OverlaySupported(root string) error {
+	return overlay.Supported(root)
+}
+
+func FuseoverlayfsSupported(root string) error {
+	return fuseoverlayfs.Supported(root)
+}
