+++ pkg/daemons/agent/agent.go
@@ -1,28 +1,23 @@
 package agent
 
 import (
-	"bufio"
 	"math/rand"
 	"os"
 	"path/filepath"
 	"strings"
 	"time"
 
-	"github.com/containerd/cgroups"
-	cgroupsv2 "github.com/containerd/cgroups/v2"
 	"github.com/opencontainers/runc/libcontainer/system"
+	"github.com/rancher/k3s/pkg/cgroups"
 	"github.com/rancher/k3s/pkg/daemons/config"
 	"github.com/rancher/k3s/pkg/daemons/executor"
 	"github.com/rancher/k3s/pkg/util"
-	"github.com/rancher/k3s/pkg/version"
 	"github.com/sirupsen/logrus"
-	"golang.org/x/sys/unix"
 	"k8s.io/apimachinery/pkg/util/net"
 	"k8s.io/component-base/logs"
-	"k8s.io/kubernetes/pkg/kubeapiserver/authorizer/modes"
-
 	_ "k8s.io/component-base/metrics/prometheus/restclient" // for client metric registration
 	_ "k8s.io/component-base/metrics/prometheus/version"    // for version metric registration
+	"k8s.io/kubernetes/pkg/kubeapiserver/authorizer/modes"
 )
 
 const unixPrefix = "unix://"
@@ -133,7 +128,7 @@ func startKubelet(cfg *config.Agent) error {
 	if err != nil || defaultIP.String() != cfg.NodeIP {
 		argsMap["node-ip"] = cfg.NodeIP
 	}
-	kubeletRoot, runtimeRoot, hasCFS, hasPIDs := CheckCgroups()
+	kubeletRoot, runtimeRoot, hasCFS, hasPIDs := cgroups.CheckCgroups()
 	if !hasCFS {
 		logrus.Warn("Disabling CPU quotas due to missing cpu.cfs_period_us")
 		argsMap["cpu-cfs-quota"] = "false"
@@ -170,20 +165,7 @@ func startKubelet(cfg *config.Agent) error {
 	}
 
 	if cfg.Rootless {
-		// "/sys/fs/cgroup" is namespaced
-		cgroupfsWritable := unix.Access("/sys/fs/cgroup", unix.W_OK) == nil
-		if hasCFS && hasPIDs && cgroupfsWritable {
-			logrus.Info("cgroup v2 controllers are delegated for rootless.")
-			// cgroupfs v2, delegated for rootless by systemd
-			argsMap["cgroup-driver"] = "cgroupfs"
-		} else {
-			logrus.Warn("cgroup v2 controllers are not delegated for rootless. Setting cgroup driver to \"none\".")
-			// flags are from https://github.com/rootless-containers/usernetes/blob/v20190826.0/boot/kubelet.sh
-			argsMap["cgroup-driver"] = "none"
-			argsMap["feature-gates=SupportNoneCgroupDriver"] = "true"
-			argsMap["cgroups-per-qos"] = "false"
-			argsMap["enforce-node-allocatable"] = ""
-		}
+		createRootlessConfig(argsMap, hasCFS, hasCFS)
 	}
 
 	if cfg.ProtectKernelDefaults {
@@ -215,112 +197,4 @@ func ImageCredProvAvailable(cfg *config.Agent) bool {
 		return false
 	}
 	return true
-}
-
-func CheckCgroups() (kubeletRoot, runtimeRoot string, hasCFS, hasPIDs bool) {
-	cgroupsModeV2 := cgroups.Mode() == cgroups.Unified
-
-	// For Unified (v2) cgroups we can directly check to see what controllers are mounted
-	// under the unified hierarchy.
-	if cgroupsModeV2 {
-		m, err := cgroupsv2.LoadManager("/sys/fs/cgroup", "/")
-		if err != nil {
-			return "", "", false, false
-		}
-		controllers, err := m.Controllers()
-		if err != nil {
-			return "", "", false, false
-		}
-		// Intentionally using an expressionless switch to match the logic below
-		for _, controller := range controllers {
-			switch {
-			case controller == "cpu":
-				hasCFS = true
-			case controller == "pids":
-				hasPIDs = true
-			}
-		}
-	}
-
-	f, err := os.Open("/proc/self/cgroup")
-	if err != nil {
-		return "", "", false, false
-	}
-	defer f.Close()
-
-	scan := bufio.NewScanner(f)
-	for scan.Scan() {
-		parts := strings.Split(scan.Text(), ":")
-		if len(parts) < 3 {
-			continue
-		}
-		controllers := strings.Split(parts[1], ",")
-		// For v1 or hybrid, controller can be a single value {"blkio"}, or a comounted set {"cpu","cpuacct"}
-		// For v2, controllers = {""} (only contains a single empty string)
-		for _, controller := range controllers {
-			switch {
-			case controller == "name=systemd" || cgroupsModeV2:
-				// If we detect that we are running under a `.scope` unit with systemd
-				// we can assume we are being directly invoked from the command line
-				// and thus need to set our kubelet root to something out of the context
-				// of `/user.slice` to ensure that `CPUAccounting` and `MemoryAccounting`
-				// are enabled, as they are generally disabled by default for `user.slice`
-				// Note that we are not setting the `runtimeRoot` as if we are running with
-				// `--docker`, we will inadvertently move the cgroup `dockerd` lives in
-				//  which is not ideal and causes dockerd to become unmanageable by systemd.
-				last := parts[len(parts)-1]
-				i := strings.LastIndex(last, ".scope")
-				if i > 0 {
-					kubeletRoot = "/" + version.Program
-				}
-			case controller == "cpu":
-				// It is common for this to show up multiple times in /sys/fs/cgroup if the controllers are comounted:
-				// as "cpu" and "cpuacct", symlinked to the actual hierarchy at "cpu,cpuacct". Unfortunately the order
-				// listed in /proc/self/cgroups may not be the same order used in /sys/fs/cgroup, so this check
-				// can fail if we use the comma-separated name. Instead, we check for the controller using the symlink.
-				p := filepath.Join("/sys/fs/cgroup", controller, parts[2], "cpu.cfs_period_us")
-				if _, err := os.Stat(p); err == nil {
-					hasCFS = true
-				}
-			case controller == "pids":
-				hasPIDs = true
-			}
-		}
-	}
-
-	// If we're running with v1 and didn't find a scope assigned by systemd, we need to create our own root cgroup to avoid
-	// just inheriting from the parent process. The kubelet will take care of moving us into it when we start it up later.
-	if kubeletRoot == "" {
-		// Examine process ID 1 to see if there is a cgroup assigned to it.
-		// When we are not in a container, process 1 is likely to be systemd or some other service manager.
-		// It either lives at `/` or `/init.scope` according to https://man7.org/linux/man-pages/man7/systemd.special.7.html
-		// When containerized, process 1 will be generally be in a cgroup, otherwise, we may be running in
-		// a host PID scenario but we don't support this.
-		g, err := os.Open("/proc/1/cgroup")
-		if err != nil {
-			return "", "", false, false
-		}
-		defer g.Close()
-		scan = bufio.NewScanner(g)
-		for scan.Scan() {
-			parts := strings.Split(scan.Text(), ":")
-			if len(parts) < 3 {
-				continue
-			}
-			controllers := strings.Split(parts[1], ",")
-			// For v1 or hybrid, controller can be a single value {"blkio"}, or a comounted set {"cpu","cpuacct"}
-			// For v2, controllers = {""} (only contains a single empty string)
-			for _, controller := range controllers {
-				switch {
-				case controller == "name=systemd" || cgroupsModeV2:
-					last := parts[len(parts)-1]
-					if last != "/" && last != "/init.scope" {
-						kubeletRoot = "/" + version.Program
-						runtimeRoot = "/" + version.Program
-					}
-				}
-			}
-		}
-	}
-	return kubeletRoot, runtimeRoot, hasCFS, hasPIDs
 }
