+++ pkg/daemons/agent/agent_windows.go
@@ -1,0 +1,6 @@
+// +build windows
+
+package agent
+
+func createRootlessConfig(argsMap map[string]string, hasCfs, hasPIDs bool) {
+}
